requirements.txt
streamlit
GitPython
esprima
astor
langchain-core
langchain-openai
openai
pydantic
tiktoken
python-dotenv
requests
groq
chromadb
sentence-transformers